fix: split infra types
