+++ src/types/infrastructure/api.d.ts
@@ -1,0 +1,11 @@
+import type { Api } from "@/shared/utils/swagger/api";
+
+declare global {
+  namespace Infra {
+    type API = Api;
+  }
+
+  interface Infra {
+    api: Infra.API;
+  }
+}
