+++ src/types/infrastructure/config.d.ts
@@ -1,0 +1,15 @@
+namespace Infra {
+  interface Config {
+    api: {
+      baseUrl: string;
+    };
+    bitcoin: {
+      url: string;
+      network: string;
+    };
+  }
+}
+
+interface Infra {
+  config: Infra.Config;
+}
