+++ src/types/infrastructure/db.d.ts
@@ -1,0 +1,23 @@
+/* eslint-disable @typescript-eslint/no-empty-object-type */
+import type { DBSchema, IDBPDatabase } from "idb";
+
+declare global {
+  namespace Infra {
+    interface Schema extends DBSchema {
+      delegations: {
+        key: "id";
+        value: {};
+      };
+      finalityProviders: {
+        key: "id";
+        value: {};
+      };
+    }
+
+    type DB = IDBPDatabase<Schema>;
+  }
+
+  interface Infra {
+    db: Infra.DB;
+  }
+}
